split PhysicsParticles into files
