+++ src/particles/PhysicsParticles.hpp
@@ -6,216 +6,19 @@
 #include <memory>
 #include <string>
 
-#include "AMReX_AmrParticles.H"
-#include "AMReX_Array.H"
 #include "AMReX_Array4.H"
-#include "AMReX_Extension.H"
 #include "AMReX_MultiFab.H"
-#include "AMReX_ParIter.H"
 #include "AMReX_ParticleInterpolators.H"
 #include "AMReX_REAL.H"
 #include "AMReX_SPACE.H"
 #include "AMReX_Vector.H"
-#include "hydro/hydro_system.hpp"
+#include "particle_creation.hpp"
+#include "particle_deposition.hpp"
+#include "particle_types.hpp"
 #include "physics_info.hpp"
-
-// Function to create bit flags: bitflag(position) = 2^(position - 1)
-// Example: bitflag<1>() = 1, bitflag<2>() = 2, bitflag<3>() = 4, ...
-template <unsigned int position>
-constexpr auto bitflag() -> unsigned int
-{
-	return 1U << (position - 1U);
-}
-
-// Particle type flags that can be combined using bitwise OR operation (|).
-// Example: To enable both CIC and Rad particles, use:
-//   particle_switch = ParticleSwitch::CIC | ParticleSwitch::Rad  (= 0b00000011)
-// To check if CIC particles are enabled:
-//   if (particle_switch & ParticleSwitch::CIC) { ... }
-enum class ParticleSwitch : unsigned int {
-	None = 0U,	    // No particles, = 0b0000
-	CIC = bitflag<1>(),   // Cloud-In-Cell (gravitating) particles, = 0b0001
-	Rad = bitflag<2>(),   // Radiation particles, = 0b0010
-	CICRad = bitflag<3>() // Combined gravitating-radiating particles, = 0b0100
-};
-
-// Enable bitwise operations on the enum class
-constexpr auto operator|(ParticleSwitch a, ParticleSwitch b) -> ParticleSwitch
-{
-	return static_cast<ParticleSwitch>(static_cast<unsigned int>(a) | static_cast<unsigned int>(b));
-}
-
-constexpr auto operator&(ParticleSwitch flags, ParticleSwitch flag) -> bool
-{
-	return (static_cast<unsigned int>(flags) & static_cast<unsigned int>(flag)) != 0;
-}
-
-// This struct should be specialized by the user application code to configure particle behavior.
-// The particle_switch member determines which particle types are enabled using bitwise flags.
-// Examples:
-// - static constexpr ParticleSwitch particle_switch = ParticleSwitch::None             -> No particles enabled
-// - static constexpr ParticleSwitch particle_switch = ParticleSwitch::CIC              -> Only CIC particles
-// - static constexpr ParticleSwitch particle_switch = ParticleSwitch::CIC | ParticleSwitch::Rad -> Both CIC and Rad particles
-// Examples that will cause a compile error:
-// - static constexpr int particle_switch = 1;
-// enum class TestEnum : unsigned int {
-// 	MISTAKE = 0b00000100U,
-// };
-// - static constexpr TestEnum particle_switch = TestEnum::MISTAKE;
-// - static constexpr ParticleSwitch particle_switch = ParticleSwitch::CIC | TestEnum::MISTAKE;
-template <typename problem_t> struct Particle_Traits {
-	static constexpr ParticleSwitch particle_switch = ParticleSwitch::None; // Determines which particle types are enabled using bitwise flags.
-};
-
-// Static assertion helper to verify that particle_switch is of the correct type
-namespace detail
-{
-template <typename problem_t> constexpr void verify_particle_switch_type()
-{
-	// This will fail to compile if particle_switch is not of type ParticleSwitch
-	static_assert(std::is_same_v<decltype(Particle_Traits<problem_t>::particle_switch), const ParticleSwitch>,
-		      "ERROR: Particle_Traits::particle_switch must be of type ParticleSwitch. "
-		      "Use any of the members of ParticleSwitch enum class, or combinations with '|'");
-}
-} // namespace detail
-
-// Assumptions for any particle type:
-// 1. For massive particles, velocity components start after mass
-// 2. Birth time, if existing, is always followed by death time
 
 namespace quokka
 {
-
-// Enum class to identify different particle types
-enum class ParticleType {
-	Rad,   // Radiation particles
-	CIC,   // Gravitating particles
-	CICRad // Gravitating radiation particles
-};
-
-//-------------------- Radiation particles --------------------
-
-// Indices for radiation particles (Rad_particles), birth time + death time + radiation groups
-enum RadParticleDataIdx {
-	RadParticleBirthTimeIdx = 0, // Time when particle becomes active
-	RadParticleDeathTimeIdx,     // Time when particle becomes inactive
-	RadParticleLumIdx	     // Base index for luminosity components
-};
-
-// Number of real components for Rad_particles, birth time + death time + radiation groups
-template <typename problem_t>
-constexpr int RadParticleRealComps = []() constexpr {
-	if constexpr (Physics_Traits<problem_t>::is_hydro_enabled || Physics_Traits<problem_t>::is_radiation_enabled) {
-		return 2 + Physics_Traits<problem_t>::nGroups; // birth_time death_time lum1 ... lumN
-	} else {
-		return 2; // birth_time death_time
-	}
-}();
-
-// Type definitions for Rad_particles container and iterator
-template <typename problem_t> using RadParticleContainer = amrex::AmrParticleContainer<RadParticleRealComps<problem_t>>;
-template <typename problem_t> using RadParticleIterator = amrex::ParIter<RadParticleRealComps<problem_t>>;
-
-//-------------------- Radiation depositions --------------------
-
-// Functor for depositing radiation energy from particles onto the grid
-struct RadDeposition {
-	double current_time{}; // Current simulation time
-	int start_part_comp{}; // Starting component in particle data
-	int start_mesh_comp{}; // Starting component in mesh data
-	int num_comp{};	       // Number of components to deposit
-	int birthTimeIndex{};  // Index for particle birth time
-
-	// Operator to perform radiation deposition using linear interpolation
-	template <typename ContainerType>
-	AMREX_GPU_DEVICE AMREX_FORCE_INLINE void operator()(const ContainerType &p, amrex::Array4<amrex::Real> const &radEnergySource,
-							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
-							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dxi) const noexcept
-	{
-		amrex::ParticleInterpolator::Linear interp(p, plo, dxi);
-		// Deposit radiation energy only if particle is active
-		interp.ParticleToMesh(p, radEnergySource, start_part_comp, start_mesh_comp, num_comp,
-				      [=] AMREX_GPU_DEVICE(const ContainerType &part, int comp) {
-					      if (current_time < part.rdata(birthTimeIndex) || current_time >= part.rdata(birthTimeIndex + 1)) {
-						      return 0.0;
-					      }
-					      return part.rdata(comp) * (AMREX_D_TERM(dxi[0], *dxi[1], *dxi[2]));
-				      });
-	}
-};
-
-//-------------------- Gravitating particles --------------------
-
-#if AMREX_SPACEDIM == 3
-
-// Indices for gravitating particles (CIC_particles), mass + 3 velocity components
-enum CICParticleDataIdx {
-	CICParticleMassIdx = 0, // Mass of the particle
-	CICParticleVxIdx,	// Velocity in x direction
-	CICParticleVyIdx,	// Velocity in y direction
-	CICParticleVzIdx	// Velocity in z direction
-};
-
-// Number of real components for CIC_particles, mass + 3 velocity components
-constexpr int CICParticleRealComps = 4;
-
-// Type definitions for CIC_particles container and iterator
-using CICParticleContainer = amrex::AmrParticleContainer<CICParticleRealComps>;
-using CICParticleIterator = amrex::ParIter<CICParticleRealComps>;
-
-//-------------------- Gravitating radiation particles --------------------
-
-// Indices for gravitating radiation particles (CICRad_particles), mass + 3 velocity components + birth time + death time + radiation groups
-enum CICRadParticleDataIdx {
-	CICRadParticleMassIdx = 0,  // Mass of the particle
-	CICRadParticleVxIdx,	    // Velocity in x direction
-	CICRadParticleVyIdx,	    // Velocity in y direction
-	CICRadParticleVzIdx,	    // Velocity in z direction
-	CICRadParticleBirthTimeIdx, // Time when particle becomes active
-	CICRadParticleDeathTimeIdx, // Time when particle becomes inactive
-	CICRadParticleLumIdx	    // Base index for luminosity components
-};
-
-// Number of real components for CICRad_particles, mass + 3 velocity components + birth time + death time + radiation groups
-template <typename problem_t>
-constexpr int CICRadParticleRealComps = []() constexpr {
-	if constexpr (Physics_Traits<problem_t>::is_hydro_enabled || Physics_Traits<problem_t>::is_radiation_enabled) {
-		return 6 + Physics_Traits<problem_t>::nGroups; // mass, vx, vy, vz, birth_time, death_time, lum[nGroups]
-	} else {
-		return 6; // mass, vx, vy, vz, birth_time, death_time
-	}
-}();
-
-// Type definitions for CICRad_particles container and iterator
-template <typename problem_t> using CICRadParticleContainer = amrex::AmrParticleContainer<CICRadParticleRealComps<problem_t>>;
-template <typename problem_t> using CICRadParticleIterator = amrex::ParIter<CICRadParticleRealComps<problem_t>>;
-
-//-------------------- Mass depositions --------------------
-
-// Functor for depositing particle mass onto the grid
-struct MassDeposition {
-	amrex::Real Gconst{};  // Gravitational constant
-	int start_part_comp{}; // Starting component in particle data
-	int start_mesh_comp{}; // Starting component in mesh data
-	int num_comp{};	       // Number of components to deposit
-
-	// Operator to perform mass deposition using linear interpolation
-	template <typename ContainerType>
-	AMREX_GPU_DEVICE AMREX_FORCE_INLINE void operator()(const ContainerType &p, amrex::Array4<amrex::Real> const &rho,
-							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
-							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dxi) const noexcept
-	{
-		amrex::ParticleInterpolator::Linear interp(p, plo, dxi);
-		// Deposit mass weighted by 4 pi G
-		interp.ParticleToMesh(p, rho, start_part_comp, start_mesh_comp, num_comp, [=] AMREX_GPU_DEVICE(const ContainerType &part, int comp) {
-			return 4.0 * M_PI * Gconst * part.rdata(comp) * (AMREX_D_TERM(dxi[0], *dxi[1], *dxi[2]));
-		});
-	}
-};
-
-#endif // AMREX_SPACEDIM == 3
-
-//-------------------- Particle depositions --------------------
 
 // Forward declarations
 template <typename problem_t> class PhysicsParticleRegister;
@@ -271,261 +74,6 @@ class PhysicsParticleDescriptorBase
 					      amrex::Real param2) const = 0;
 #endif // AMREX_SPACEDIM == 3
 };
-
-// Helper namespace with implementation details for particle creation
-namespace ParticleCreationImpl
-{
-// Common implementation of particle creation logic
-template <typename problem_t, typename ContainerType, template <typename> class CheckerType, template <typename> class CreatorType>
-static void createParticlesImpl(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
-				amrex::Real param1, amrex::Real param2)
-{
-	if (container != nullptr) {
-		if (mass_idx >= 0 && mass_idx + 3 < ContainerType::ParticleType::NReal) {
-			// Use the provided ParticleChecker type
-			CheckerType<problem_t> particle_checker(param1, param2);
-
-			for (amrex::MFIter mfi = container->MakeMFIter(lev); mfi.isValid(); ++mfi) {
-				const auto &box = mfi.validbox();
-				const auto &state_arr = state.array(mfi);
-				const auto &geom = container->Geom(lev);
-				const auto dx = geom.CellSizeArray();
-				const auto plo = geom.ProbLoArray();
-
-				// Count particles to be created in this box
-				amrex::Gpu::DeviceVector<unsigned int> counts(box.numPts()); // 1 if cell creates particle, 0 if not
-				amrex::Gpu::DeviceVector<unsigned int> offset(box.numPts()); // Will store starting index for each cell's particle
-				auto *pcounts = counts.data();
-
-				// Count potential particles per cell
-				amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
-					const amrex::IntVect iv(AMREX_D_DECL(i, j, k));
-					const auto index = box.index(iv);
-					// Check if we should create a particle at this location and time
-					pcounts[index] = particle_checker(state_arr, i, j, k, dx, current_time, dt) ? 1 : 0; // NOLINT
-				});
-
-				// Calculate exclusive prefix sum to get unique position for each particle
-				// Example: counts  = [1, 0, 1, 0, 1]
-				//         offset  = [0, 1, 1, 2, 2]
-				const unsigned int max_new_particles = amrex::Scan::ExclusiveSum(counts.size(), counts.data(), offset.data());
-
-				// Update NextID to include particles that will be created
-				const amrex::Long pid = ContainerType::ParticleType::NextID();
-				ContainerType::ParticleType::NextID(pid + max_new_particles);
-
-				// Get the particle tile and prepare for new particles
-				auto &particle_tile = container->DefineAndReturnParticleTile(lev, mfi);
-				auto &aos = particle_tile.GetArrayOfStructs();
-				const int old_size = aos.size();
-				aos.resize(old_size + max_new_particles);
-
-				// Create the particles
-				auto *poffset = offset.data();
-				auto *pdata = aos.data() + old_size;
-				const int cpu_id = amrex::ParallelDescriptor::MyProc();
-
-				// Initialize particle creator functor using the provided ParticleCreator type
-				CreatorType<problem_t> particle_creator(mass_idx, cpu_id, pid, param1, param2);
-
-				amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
-					const amrex::IntVect iv(AMREX_D_DECL(i, j, k));
-					const auto index = box.index(iv);
-
-					if (pcounts[index] > 0) {						  // NOLINT
-						auto &p = pdata[poffset[index]];				  // NOLINT
-						particle_creator(p, state_arr, i, j, k, dx, plo, poffset[index]); // NOLINT
-					}
-				});
-			}
-		}
-	}
-}
-} // namespace ParticleCreationImpl
-
-// Traits class for specializing particle creation behavior
-template <ParticleType particleType> struct ParticleCreationTraits {
-	// Default nested ParticleChecker - determines if a particle should be created at a location
-	template <typename problem_t> struct ParticleChecker {
-		double param1;
-		double param2;
-		AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
-
-		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
-		{
-			// Default implementation creates no particles
-			amrex::ignore_unused(state_arr, i, j, k, dx, current_time, dt, param1, param2);
-			return false;
-		}
-	};
-
-	// Default nested ParticleCreator - initializes a particle's properties
-	template <typename problem_t> struct ParticleCreator {
-		int mass_idx;
-		int cpu_id;
-		amrex::Long pid_start;
-		amrex::Real param1;
-		amrex::Real param2;
-
-		AMREX_GPU_HOST_DEVICE
-		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
-		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
-		{
-		}
-
-		template <typename ParticleType, typename StateArray>
-		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
-		{
-			// Default implementation does nothing
-			amrex::ignore_unused(p, state_arr, i, j, k, dx, plo, particle_offset);
-		}
-	};
-
-	// Main method to create particles - uses the helper implementation
-	template <typename problem_t, typename ContainerType>
-	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
-				    amrex::Real param1, amrex::Real param2)
-	{
-		// Use the common implementation with our checker and creator types
-		ParticleCreationImpl::createParticlesImpl<problem_t, ContainerType, ParticleCreationTraits<particleType>::template ParticleChecker,
-							  ParticleCreationTraits<particleType>::template ParticleCreator>(container, mass_idx, state, lev,
-															  current_time, dt, param1, param2);
-	}
-};
-
-// Specialization for CIC particles
-template <> struct ParticleCreationTraits<ParticleType::CIC> {
-	// Specialized nested ParticleChecker for CIC particles
-	template <typename problem_t> struct ParticleChecker {
-		double param1;
-		double param2;
-		AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
-
-		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
-		{
-			// A simple demonstration of particle creation
-			// Could check density threshold or other state-based conditions
-			amrex::ignore_unused(state_arr, dx);
-			const int spacing = 16;
-			const bool is_create_particle_1 = current_time <= param1 && current_time + dt > param1;
-			const bool is_create_particle_2 = current_time <= param2 && current_time + dt > param2;
-			return (is_create_particle_1 || is_create_particle_2) && (i != 0 && i % spacing == 0) && (j != 0 && j % spacing == 0) &&
-			       (k != 0 && k % spacing == 0);
-		}
-	};
-
-	// Specialized nested ParticleCreator for CIC particles
-	template <typename problem_t> struct ParticleCreator {
-		int mass_idx;
-		int cpu_id;
-		amrex::Long pid_start;
-		amrex::Real param1;
-		amrex::Real param2;
-
-		AMREX_GPU_HOST_DEVICE
-		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
-		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
-		{
-		}
-
-		template <typename ParticleType, typename StateArray>
-		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
-		{
-			// A simple demonstration of particle creation
-			if (mass_idx + 3 < ParticleType::NReal) {
-				p.pos(0) = plo[0] + (i + 0.5) * dx[0];
-				p.pos(1) = plo[1] + (j + 0.5) * dx[1];
-				p.pos(2) = plo[2] + (k + 0.5) * dx[2];
-
-				// Set particle ID and CPU
-				p.id() = pid_start + particle_offset;
-				p.cpu() = cpu_id;
-
-				// Set particle mass and velocities
-				const amrex::Real cell_volume = AMREX_D_TERM(dx[0], *dx[1], *dx[2]);
-				const amrex::Real cell_density = state_arr(i, j, k, HydroSystem<problem_t>::density_index);
-				const amrex::Real cell_mass = cell_density * cell_volume;
-
-				// Initialize particle properties
-				p.rdata(mass_idx) = 0.5 * cell_mass;
-				p.rdata(mass_idx + 1) = state_arr(i, j, k, HydroSystem<problem_t>::x1Momentum_index) / cell_density;
-				p.rdata(mass_idx + 2) = state_arr(i, j, k, HydroSystem<problem_t>::x2Momentum_index) / cell_density;
-				p.rdata(mass_idx + 3) = state_arr(i, j, k, HydroSystem<problem_t>::x3Momentum_index) / cell_density;
-
-				// Update cell density (remove mass that was given to particle)
-				state_arr(i, j, k, HydroSystem<problem_t>::density_index) = 0.5 * cell_density;
-			}
-		}
-	};
-
-	// Main method to create particles - uses the helper implementation
-	template <typename problem_t, typename ContainerType>
-	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
-				    amrex::Real param1, amrex::Real param2)
-	{
-		// Use the common implementation with our checker and creator types
-		ParticleCreationImpl::createParticlesImpl<problem_t, ContainerType, ParticleCreationTraits<ParticleType::CIC>::template ParticleChecker,
-							  ParticleCreationTraits<ParticleType::CIC>::template ParticleCreator>(
-		    container, mass_idx, state, lev, current_time, dt, param1, param2);
-	}
-};
-
-// // Example: specialization for Radiation particles
-// template <> struct ParticleCreationTraits<ParticleType::Rad> {
-// 	// Specialized nested ParticleChecker for Radiation particles
-// 	template <typename problem_t> struct ParticleChecker {
-// 		double param1;
-// 		double param2;
-// 		AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
-
-// 		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
-// 						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
-// 		{
-// 			// To be implemented
-// 			return true;
-// 		}
-// 	};
-
-// 	// Specialized nested ParticleCreator for Radiation particles
-// 	template <typename problem_t> struct ParticleCreator {
-// 		int mass_idx;
-// 		int cpu_id;
-// 		amrex::Long pid_start;
-// 		amrex::Real param1;
-// 		amrex::Real param2;
-
-// 		AMREX_GPU_HOST_DEVICE
-// 		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
-// 		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
-// 		{
-// 		}
-
-// 		template <typename ParticleType, typename StateArray>
-// 		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-// 						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
-// 						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
-// 		{
-// 			// To be implemented
-// 		}
-// 	};
-
-// 	// Main method to create particles - uses the helper implementation
-// 	template <typename problem_t, typename ContainerType>
-// 	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
-// 				    amrex::Real param1, amrex::Real param2)
-// 	{
-// 		// Use the common implementation with our checker and creator types
-// 		ParticleCreationImpl::createParticlesImpl<problem_t, ContainerType, ParticleCreationTraits<ParticleType::Rad>::template ParticleChecker,
-// 							  ParticleCreationTraits<ParticleType::Rad>::template ParticleCreator>(container, mass_idx, state, lev,
-// 															  current_time, dt, param1, param2);
-// 	}
-// };
 
 // Concrete implementation of particle descriptor for specific container types
 template <typename ContainerType, typename problem_t, ParticleType particleType> class PhysicsParticleDescriptor : public PhysicsParticleDescriptorBase
