+++ src/particles/particle_deposition.hpp
@@ -1,0 +1,69 @@
+#ifndef PARTICLE_DEPOSITION_HPP_
+#define PARTICLE_DEPOSITION_HPP_
+
+#include "AMReX_Array4.H"
+#include "AMReX_Extension.H"
+#include "AMReX_ParticleInterpolators.H"
+#include "AMReX_REAL.H"
+
+namespace quokka
+{
+
+//-------------------- Radiation depositions --------------------
+
+// Functor for depositing radiation energy from particles onto the grid
+struct RadDeposition {
+	double current_time{}; // Current simulation time
+	int start_part_comp{}; // Starting component in particle data
+	int start_mesh_comp{}; // Starting component in mesh data
+	int num_comp{};	       // Number of components to deposit
+	int birthTimeIndex{};  // Index for particle birth time
+
+	// Operator to perform radiation deposition using linear interpolation
+	template <typename ContainerType>
+	AMREX_GPU_DEVICE AMREX_FORCE_INLINE void operator()(const ContainerType &p, amrex::Array4<amrex::Real> const &radEnergySource,
+							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
+							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dxi) const noexcept
+	{
+		amrex::ParticleInterpolator::Linear interp(p, plo, dxi);
+		// Deposit radiation energy only if particle is active
+		interp.ParticleToMesh(p, radEnergySource, start_part_comp, start_mesh_comp, num_comp,
+				      [=] AMREX_GPU_DEVICE(const ContainerType &part, int comp) {
+					      if (current_time < part.rdata(birthTimeIndex) || current_time >= part.rdata(birthTimeIndex + 1)) {
+						      return 0.0;
+					      }
+					      return part.rdata(comp) * (AMREX_D_TERM(dxi[0], *dxi[1], *dxi[2]));
+				      });
+	}
+};
+
+#if AMREX_SPACEDIM == 3
+
+//-------------------- Mass depositions --------------------
+
+// Functor for depositing particle mass onto the grid
+struct MassDeposition {
+	amrex::Real Gconst{};  // Gravitational constant
+	int start_part_comp{}; // Starting component in particle data
+	int start_mesh_comp{}; // Starting component in mesh data
+	int num_comp{};	       // Number of components to deposit
+
+	// Operator to perform mass deposition using linear interpolation
+	template <typename ContainerType>
+	AMREX_GPU_DEVICE AMREX_FORCE_INLINE void operator()(const ContainerType &p, amrex::Array4<amrex::Real> const &rho,
+							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
+							    amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dxi) const noexcept
+	{
+		amrex::ParticleInterpolator::Linear interp(p, plo, dxi);
+		// Deposit mass weighted by 4 pi G
+		interp.ParticleToMesh(p, rho, start_part_comp, start_mesh_comp, num_comp, [=] AMREX_GPU_DEVICE(const ContainerType &part, int comp) {
+			return 4.0 * M_PI * Gconst * part.rdata(comp) * (AMREX_D_TERM(dxi[0], *dxi[1], *dxi[2]));
+		});
+	}
+};
+
+#endif // AMREX_SPACEDIM == 3
+
+} // namespace quokka
+
+#endif // PARTICLE_DEPOSITION_HPP_
